% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnem_bip39.R
\name{bip39_normalize}
\alias{bip39_normalize}
\title{Normalize UTF-8 and strip whitespace}
\usage{
bip39_normalize(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Normalize UTF-8 and strip whitespace
}
\keyword{internal}
