% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbkdf2.r
\name{f_PBKDF2}
\alias{f_PBKDF2}
\title{f function}
\usage{
f_PBKDF2(passphrase, salt, iterations, prf, index)
}
\arguments{
\item{passphrase}{String to be expanded into a key.}

\item{salt}{Raw bytes to expand the set of possible keys.}

\item{iterations}{Number of times to apply the pseudorandom function}

\item{prf}{Pseudorandom function()}

\item{index}{the block index to append to the salt in the first prf call}
}
\value{
It returns a (typically fixed length) raw vector.
}
\description{
The internal function f in the PBKDF2 algorithm.  Returns a block of hlen
key bytes, where hlen is the block size of the underlying pseudorandom
function, prf().  The pseudorandom function takes a passphrase, a
variable-length raw vector; and a salt, also a variable-length raw vector.
}
\keyword{internal}
