% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdf_hkdf.R
\name{hkdf_expand}
\alias{hkdf_expand}
\title{HKDF - Expand}
\usage{
hkdf_expand(prk, info, len, hash)
}
\arguments{
\item{prk}{raw vector}

\item{info}{raw vector}

\item{len}{positive integer}

\item{hash}{hmac-hash function}
}
\value{
raw vector
}
\description{
Internal invariant:
\itemize{
\item \code{prk}, \code{info} are raw vectors
\item \code{len} is a positive integer <= 255*HashLen
\item \code{hash} is a function(key, data) -> raw
}
}
\keyword{internal}
