% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdf_hkdf.R
\name{hkdf_extract}
\alias{hkdf_extract}
\title{HKDF - Extract}
\usage{
hkdf_extract(salt, ikm, hash)
}
\arguments{
\item{salt}{raw vector}

\item{ikm}{raw vector}

\item{hash}{hmac-hash function}
}
\value{
raw vector
}
\description{
Internal invariant:
\itemize{
\item \code{salt} and \code{ikm} are raw vectors
\item \code{hash} is a function(key, data) -> raw
\item Validation is guaranteed by the public wrapper
}
}
\keyword{internal}
