% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudorandom_functions.R
\name{HMAC_SHA1}
\alias{HMAC_SHA1}
\title{HMAC-SHA1}
\usage{
HMAC_SHA1(key, object)
}
\arguments{
\item{key}{HMAC key}

\item{object}{object ro be hashed}
}
\description{
Pseudorandom function to apply HMAC with SHA1 hash algorithm
}
