% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudorandom_functions.R
\name{HMAC_SHA2}
\alias{HMAC_SHA2}
\alias{HMAC_SHA224}
\alias{HMAC_SHA256}
\alias{HMAC_SHA384}
\alias{HMAC_SHA512}
\title{HMAC-SHA2}
\usage{
HMAC_SHA224(key, object)

HMAC_SHA256(key, object)

HMAC_SHA384(key, object)

HMAC_SHA512(key, object)
}
\arguments{
\item{key}{HMAC key}

\item{object}{object ro be hashed}
}
\description{
Pseudorandom functions to apply HMAC with SHA2 hash algorithms
}
