% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudorandom_functions.R
\name{HMAC_SHA3}
\alias{HMAC_SHA3}
\alias{HMAC_SHA3_224}
\alias{HMAC_SHA3_256}
\alias{HMAC_SHA3_384}
\alias{HMAC_SHA3_512}
\title{HMAC-SHA3}
\usage{
HMAC_SHA3_224(key, object)

HMAC_SHA3_256(key, object)

HMAC_SHA3_384(key, object)

HMAC_SHA3_512(key, object)
}
\arguments{
\item{key}{HMAC key}

\item{object}{object ro be hashed}
}
\description{
Pseudorandom functions to apply HMAC with SHA3 hash algorithms
}
