% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudorandom_functions.R
\name{HMAC_SHA_256}
\alias{HMAC_SHA_256}
\title{HMAC-SHA2}
\usage{
HMAC_SHA_256(key, object)
}
\arguments{
\item{key}{HMAC key}

\item{object}{object ro be hashed}
}
\description{
HMAC-SHA2
}
\keyword{internal}
