% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeStringRaw}
\alias{makeStringRaw}
\title{makeStringRaw}
\usage{
makeStringRaw(obj)
}
\arguments{
\item{obj}{the UTF-8 string to be converted}
}
\description{
Convert a UTF-8 string to raw bytes.
}
\keyword{internal}
