% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbkdf2.r
\name{PBKDF2}
\alias{PBKDF2}
\title{PBKDF2}
\usage{
PBKDF2(passphrase, salt, dkLen, iterations = 1000, prf = HMAC_SHA2_256)
}
\arguments{
\item{passphrase}{String to be expanded into a key.}

\item{salt}{Raw bytes to expand the set of possible keys.}

\item{dkLen}{Number of key bytes to return}

\item{iterations}{Number of times to apply the pseudorandom function when
computing a key block in the function f (default: 1000)}

\item{prf}{Pseudorandom function (default: hmac-sha2-256)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A `pbkdf2_key' object including the key of requested
               length and metadata about key generation.
}\if{html}{\out{</div>}}
}
\description{
Get the requested number of bytes from the expanded key.
}
\examples{
key <- PBKDF2("pass", "salt", 32)

}
