% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnem_bip39.R
\name{rkdf_mnemonic_to_words}
\alias{rkdf_mnemonic_to_words}
\title{Entropy to Mnemonic Words}
\usage{
rkdf_mnemonic_to_words(entropy)
}
\arguments{
\item{entropy}{raw vector of 16, 20, 24, 28, or 32 bytes (128, 160, 192, 224, 256 bits)}
}
\value{
character vector of words
}
\description{
Convert random bytes to a BIP39 mnemonic sentence.
}
\examples{
entropy <- openssl::rand_bytes(16)
rkdf_mnemonic_to_words(entropy)

}
