% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uintToRaw}
\alias{uintToRaw}
\title{uintToRaw}
\usage{
uintToRaw(num, minLen = 4)
}
\arguments{
\item{num}{the number, which must be positive with no fractional part}

\item{minLen}{the minimum number of bytes in the result (default: 4)}
}
\description{
Convert an unsigned integer from R numeric to its big-endian binary form.
Used to convert the index argument of the f function into the required
4-byte vector to append to the salt.
}
\keyword{internal}
