% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudorandom_functions.R
\name{HMAC_SHA2}
\alias{HMAC_SHA2}
\alias{HMAC_SHA2_224}
\alias{HMAC_SHA2_256}
\alias{HMAC_SHA2_384}
\alias{HMAC_SHA2_512}
\title{HMAC-SHA2}
\usage{
HMAC_SHA2_224(key, object)

HMAC_SHA2_256(key, object)

HMAC_SHA2_384(key, object)

HMAC_SHA2_512(key, object)
}
\arguments{
\item{key}{HMAC key}

\item{object}{object ro be hashed}
}
\description{
Pseudorandom functions to apply HMAC with SHA2 hash algorithms
}
